requirements.txt
google-generativeai>=0.4.0
Pillow>=10.1.0
selenium>=4.15.1